+++ 42_Heilbronn/Basecamp/libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: jkollner <jkollner@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/13 15:17:58 by jkollner          #+#    #+#             */
-/*   Updated: 2022/11/07 12:35:38 by jkollner         ###   ########.fr       */
+/*   Updated: 2022/11/14 11:33:02 by jkollner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,19 +15,19 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	needle_len;
-	int	counter;
-	int	found_counter;
+	size_t	needle_len;
+	size_t	counter;
+	size_t	found_counter;
 
 	counter = 0;
 	needle_len = ft_strlen(needle);
 	if (*needle == '\0')
 		return ((char *)haystack);
-	if (needle == haystack && needle_len == (int)ft_strlen(haystack))
+	if (needle == haystack && needle_len == ft_strlen(haystack))
 		return ((char *)haystack);
-	if ((int)len == 0)
+	if (len == 0)
 		return (NULL);
-	while (haystack[counter] != '\0' && (counter + needle_len) <= (int)len)
+	while (haystack[counter] != '\0' && (counter + needle_len) <= len)
 	{
 		found_counter = 0;
 		while (haystack[counter + found_counter] == needle[found_counter]
